botaselole7/requirements.txt
Flask
pymongo
apscheduler
discord.py
requests
